+++ traffic_ops/traffic_ops_golang/server/servers.go
@@ -64,24 +64,25 @@ JOIN type t ON s.type = t.id
 
 /* language=SQL */
 const dssTopologiesJoinSubquery = `
-SELECT
-	td.id deliveryservice,
-	s.id "server"
+(SELECT
+	ARRAY_AGG(CAST(ROW(td.id, s.id, NULL) AS deliveryservice_server))
 FROM "server" s
 JOIN cachegroup c on s.cachegroup = c.id
 LEFT JOIN topology_cachegroup tc ON c.name = tc.cachegroup
 LEFT JOIN deliveryservice td ON td.topology = tc.topology
-UNION
+),
 `
 
 /* language=SQL */
 const deliveryServiceServersJoin = `
 FULL OUTER JOIN (
-%s
-SELECT
-	dss.deliveryservice,
-	dss."server"
-FROM deliveryservice_server dss
+SELECT (dss.dss_record).deliveryservice, (dss.dss_record).server FROM (
+	SELECT UNNEST(COALESCE(
+		%s
+		(SELECT
+			ARRAY_AGG(CAST(ROW(dss.deliveryservice, dss."server", NULL) AS deliveryservice_server))
+		FROM deliveryservice_server dss)
+	)) AS dss_record) AS dss
 ) dss ON dss.server = s.id
 JOIN deliveryservice d ON cdn.id = d.cdn_id AND dss.deliveryservice = d.id
 `
@@ -772,6 +773,7 @@ func getServers(h http.Header, params map[string]string, tx *sqlx.Tx, user *auth
 		"physLocation":     dbhelpers.WhereColumnInfo{"s.phys_location", api.IsInt},
 		"profileId":        dbhelpers.WhereColumnInfo{"s.profile", api.IsInt},
 		"status":           dbhelpers.WhereColumnInfo{"st.name", nil},
+		"topology":         dbhelpers.WhereColumnInfo{"tc.topology", nil},
 		"type":             dbhelpers.WhereColumnInfo{"t.name", nil},
 		"dsId":             dbhelpers.WhereColumnInfo{"dss.deliveryservice", nil},
 	}
@@ -813,6 +815,13 @@ func getServers(h http.Header, params map[string]string, tx *sqlx.Tx, user *auth
 		}
 		usesMids = dsType.UsesMidCache()
 		log.Debugf("Servers for ds %d; uses mids? %v\n", dsID, usesMids)
+	}
+
+	if _, ok := params[`topology`]; ok {
+		/* language=SQL */
+		queryAddition += `
+			JOIN topology_cachegroup tc ON cg."name" = tc.cachegroup
+`
 	}
 
 	where, orderBy, pagination, queryValues, errs := dbhelpers.BuildWhereAndOrderByAndPagination(params, queryParamsToSQLCols)